splits up transforms
